moved some methods around to different files and made updates to new csv.rs file
