+++ tests/csv_tests.rs
@@ -1,0 +1,4 @@
+#[test]
+fn create_new_csv_file(){
+
+}
